# web/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Job Seeker API",
    "description": "Dopasowuje oferty pracy z serwisów (na razie JustJoin.it) do CV kandydata.",
    "version": "0.2.0"
  },
  "paths": {
...
    "/offers/{offer_id}": {
      "get": {
        "tags": [
          "offers"
        ],
        "summary": "Get Offer",
        "description": "Offer from the local database (``source:external_id``), with its description fetched on demand.",
        "operationId": "get_offer_offers__offer_id__get",
        "parameters": [
          {
            "name": "offer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Offer Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobOffer"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/offers/{offer_id}/status": {
      "put": {
        "tags": [
          "offers"
        ],
        "summary": "Set Offer Status",
        "description": "Mark an offer as saved or hidden (hidden offers are left out of /matches); null clears the mark.",
        "operationId": "set_offer_status_offers__offer_id__status_put",
        "parameters": [
          {
            "name": "offer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Offer Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OfferStatusUpdate"
              }
            }
          }
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/offers/{offer_id}/assess": {
      "post": {
        "tags": [
          "offers"
        ],
        "summary": "Assess Offer",
        "description": "AI-assess one offer, e.g. one outside the top N rated in /matches?mode=ai.",
        "operationId": "assess_offer_offers__offer_id__assess_post",
        "parameters": [
          {
            "name": "offer_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Offer Id"
            }
          },
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Search"
            }
          },
          {
            "name": "provider",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Provider"
            }
          },
          {
            "name": "model",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MatchResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/profile/cv": {
      "post": {
        "tags": [
          "profile"
        ],
        "summary": "Upload Cv",
        "description": "Upload a new CV (PDF). It becomes the newest file in the CV folder and the profile is rebuilt.",
        "operationId": "upload_cv_profile_cv_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_cv_profile_cv_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/profile/overrides": {
      "get": {
        "tags": [
          "profile"
        ],
        "summary": "Get Overrides",
        "operationId": "get_overrides_profile_overrides_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileOverrides-Output"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "profile"
        ],
        "summary": "Put Overrides",
        "description": "Replace manual profile corrections (kept across CV changes) and return the resulting profile.",
        "operationId": "put_overrides_profile_overrides_put",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProfileOverrides-Input"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AIResult": {
        "properties": {
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "model": {
            "type": "string",
            "title": "Model"
          },
          "assessment": {
            "$ref": "#/components/schemas/AIAssessment"
          },
          "input_tokens": {
            "type": "integer",
            "title": "Input Tokens",
            "default": 0
          },
          "output_tokens": {
            "type": "integer",
            "title": "Output Tokens",
            "default": 0
          },
          "cached": {
            "type": "boolean",
            "title": "Cached",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "provider",
          "model",
          "assessment",
          "input_tokens",
          "output_tokens",
          "cached"
        ],
        "title": "AIResult"
      },
...
      "Body_upload_cv_profile_cv_post": {
        "properties": {
          "file": {
            "type": "string",
            "contentMediaType": "application/octet-stream",
            "title": "File"
          }
        },
        "type": "object",
        "required": [
          "file"
        ],
        "title": "Body_upload_cv_profile_cv_post"
      },
      "CandidateProfile": {
        "properties": {
          "parser_version": {
            "type": "integer",
            "title": "Parser Version",
            "description": "Version of the CV parser that built this profile",
            "default": 0
          },
          "source_file": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source File"
          },
          "source_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Source Hash",
            "description": "SHA-256 of the CV file the profile was built from"
          },
          "built_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Built At"
          },
          "headline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headline"
          },
          "location": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location"
          },
          "years_of_experience": {
            "type": "number",
            "title": "Years Of Experience",
            "default": 0
          },
          "seniority": {
            "$ref": "#/components/schemas/Seniority",
            "default": "mid"
          },
          "skills": {
            "items": {
              "$ref": "#/components/schemas/CandidateSkill"
            },
            "type": "array",
            "title": "Skills"
          },
          "skill_years": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Skill Years",
            "description": "Canonical skill -> years used in positions listed in the CV"
          },
          "languages": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Languages",
            "description": "ISO code -> CEFR level"
          },
          "cv_text": {
            "type": "string",
            "title": "Cv Text",
            "description": "CV text with contact data removed; sent to AI scorers",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "parser_version",
          "source_file",
          "source_hash",
          "built_at",
          "headline",
          "location",
          "years_of_experience",
          "seniority",
          "skills",
          "skill_years",
          "languages",
          "cv_text"
        ],
        "title": "CandidateProfile"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "JobOffer": {
        "properties": {
          "source": {
            "type": "string",
            "title": "Source"
          },
          "external_id": {
            "type": "string",
            "title": "External Id"
          },
          "url": {
            "type": "string",
            "title": "Url"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "company": {
            "type": "string",
            "title": "Company"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "seniority": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Seniority"
              },
              {
                "type": "null"
              }
            ]
          },
          "workplace_type": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WorkplaceType"
              },
              {
                "type": "null"
              }
            ]
          },
          "working_time": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Working Time"
          },
          "locations": {
            "items": {
              "$ref": "#/components/schemas/Location"
            },
            "type": "array",
            "title": "Locations"
          },
          "required_skills": {
            "items": {
              "$ref": "#/components/schemas/Skill"
            },
            "type": "array",
            "title": "Required Skills"
          },
          "nice_to_have_skills": {
            "items": {
              "$ref": "#/components/schemas/Skill"
            },
            "type": "array",
            "title": "Nice To Have Skills"
          },
          "languages": {
            "items": {
              "$ref": "#/components/schemas/LanguageRequirement"
            },
            "type": "array",
            "title": "Languages"
          },
          "salaries": {
            "items": {
              "$ref": "#/components/schemas/Salary"
            },
            "type": "array",
            "title": "Salaries"
          },
          "published_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Published At"
          },
          "expires_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Expires At"
          },
          "apply_url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Apply Url"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Plain-text description, filled by fetch_details"
          },
          "extra": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Extra",
            "description": "Source-specific identifiers, e.g. slug"
          },
          "id": {
            "type": "string",
            "title": "Id",
            "description": "Globally unique id: ``source:external_id``.",
            "readOnly": true
          }
        },
        "type": "object",
        "required": [
          "source",
          "external_id",
          "url",
          "title",
          "company",
          "category",
          "seniority",
          "workplace_type",
          "working_time",
          "locations",
          "required_skills",
          "nice_to_have_skills",
          "languages",
          "salaries",
          "published_at",
          "expires_at",
          "apply_url",
          "description",
          "extra",
          "id"
        ],
        "title": "JobOffer"
      },
      "LanguageRequirement": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code",
            "description": "ISO 639-1 language code, e.g. 'en'"
          },
          "level": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level",
            "description": "CEFR level, e.g. 'B2'"
          }
        },
        "type": "object",
        "required": [
          "code",
          "level"
        ],
        "title": "LanguageRequirement"
      },
      "Location": {
        "properties": {
          "city": {
            "type": "string",
            "title": "City"
          },
          "street": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Street"
          }
        },
        "type": "object",
        "required": [
          "city",
          "street"
        ],
        "title": "Location"
      },
...
      "MatchResult": {
        "properties": {
          "offer": {
            "$ref": "#/components/schemas/JobOffer"
          },
          "rule": {
            "$ref": "#/components/schemas/RuleScore"
          },
          "ai": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/AIResult"
              },
              {
                "type": "null"
              }
            ]
          },
          "final_score": {
            "type": "number",
            "title": "Final Score"
          },
          "status": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "saved",
                  "hidden"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Status",
            "description": "The user's mark on the offer"
          }
        },
        "type": "object",
        "required": [
          "offer",
          "rule",
          "ai",
          "final_score",
          "status"
        ],
        "title": "MatchResult"
      },
...
      "OfferStatusUpdate": {
        "properties": {
          "status": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "saved",
                  "hidden"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "OfferStatusUpdate"
      },
      "ProfileOverrides-Input": {
        "properties": {
          "add_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Add Skills",
            "description": "Added as core skills"
          },
          "remove_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Remove Skills"
          },
          "skill_weights": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Skill Weights",
            "description": "Skill -> weight 0..1"
          },
          "skill_years": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Skill Years",
            "description": "Skill -> years of commercial use, e.g. {'python': 2}"
          },
          "years_of_experience": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Years Of Experience"
          },
          "seniority": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Seniority"
              },
              {
                "type": "null"
              }
            ]
          },
          "languages": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Languages",
            "description": "ISO code -> CEFR level"
          },
          "extra_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extra Notes",
            "description": "Extra context for AI scoring, e.g. what you want"
          }
        },
        "type": "object",
        "title": "ProfileOverrides",
        "description": "Manual corrections applied on top of the profile parsed from the CV."
      },
      "ProfileOverrides-Output": {
        "properties": {
          "add_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Add Skills",
            "description": "Added as core skills"
          },
          "remove_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Remove Skills"
          },
          "skill_weights": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Skill Weights",
            "description": "Skill -> weight 0..1"
          },
          "skill_years": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Skill Years",
            "description": "Skill -> years of commercial use, e.g. {'python': 2}"
          },
          "years_of_experience": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Years Of Experience"
          },
          "seniority": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Seniority"
              },
              {
                "type": "null"
              }
            ]
          },
          "languages": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Languages",
            "description": "ISO code -> CEFR level"
          },
          "extra_notes": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Extra Notes",
            "description": "Extra context for AI scoring, e.g. what you want"
          }
        },
        "type": "object",
        "required": [
          "add_skills",
          "remove_skills",
          "skill_weights",
          "skill_years",
          "years_of_experience",
          "seniority",
          "languages",
          "extra_notes"
        ],
        "title": "ProfileOverrides",
        "description": "Manual corrections applied on top of the profile parsed from the CV."
      },
      "ProfileResponse": {
        "properties": {
          "profile": {
            "$ref": "#/components/schemas/CandidateProfile"
          },
          "profile_hash": {
            "type": "string",
            "title": "Profile Hash"
          },
          "rebuilt": {
            "type": "boolean",
            "title": "Rebuilt"
          },
          "warning": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Warning"
          }
        },
        "type": "object",
        "required": [
          "profile",
          "profile_hash",
          "rebuilt",
          "warning"
        ],
        "title": "ProfileResponse"
      },
      "RuleScore": {
        "properties": {
          "score": {
            "type": "number",
            "maximum": 100.0,
            "minimum": 0.0,
            "title": "Score"
          },
          "breakdown": {
            "additionalProperties": {
              "type": "number"
            },
            "type": "object",
            "title": "Breakdown",
            "description": "Component -> 0..1"
          },
          "matched_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Matched Skills"
          },
          "missing_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Missing Skills"
          },
          "notes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Notes"
          },
          "effective_years": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Effective Years",
            "description": "Candidate's experience in this offer's main technologies"
          },
          "effective_level": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/Seniority"
              },
              {
                "type": "null"
              }
            ]
          },
          "main_skills": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Main Skills",
            "description": "Offer skills used for effective_years"
          }
        },
        "type": "object",
        "required": [
          "score",
          "breakdown",
          "matched_skills",
          "missing_skills",
          "notes",
          "effective_years",
          "effective_level",
          "main_skills"
        ],
        "title": "RuleScore"
      },
...
      "Salary": {
        "properties": {
          "contract": {
            "type": "string",
            "title": "Contract",
            "description": "b2b, permanent, mandate_contract, internship or any"
          },
          "min_pln_month": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Min Pln Month"
          },
          "max_pln_month": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Max Pln Month"
          },
          "gross": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Gross"
          },
          "original_currency": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Currency"
          },
          "original_unit": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Unit"
          },
          "original_min": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Min"
          },
          "original_max": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Original Max"
          }
        },
        "type": "object",
        "required": [
          "contract",
          "min_pln_month",
          "max_pln_month",
          "gross",
          "original_currency",
          "original_unit",
          "original_min",
          "original_max"
        ],
        "title": "Salary",
        "description": "Salary range normalized to PLN per month."
      },
...
      "Seniority": {
        "type": "string",
        "enum": [
          "intern",
          "junior",
          "mid",
          "senior",
          "manager",
          "c_level"
        ],
        "title": "Seniority"
      },
...
      "Skill": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "level": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Level",
            "description": "Required proficiency, 1 (basic) to 5 (expert)"
          }
        },
        "type": "object",
        "required": [
          "name",
          "level"
        ],
        "title": "Skill"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
      "WorkplaceType": {
        "type": "string",
        "enum": [
          "remote",
          "hybrid",
          "office"
        ],
        "title": "WorkplaceType"
      }
    }
  }
}